// src/interfaces/categories.interface.ts
export enum CategoryLabel  {
    Salary = "Salary",
    Extra = "Extra",
    Dividends = "Dividends",
    Rentals = "Rentals",
    Sales = "Sales",
    Gifts = "Gifts",
    Gains = "Gains",
    Clients = "Clients",
    Returns = "Returns",
    Scholarships = "Scholarships",
    Interest = "Interest",
    Food = "Food",
    Groceries = "Groceries",
    Transport = "Transport",
    Home = "Home",
    Health = "Health",
    Education = "Education",
    Gaming = "Gaming",
    Travel = "Travel",
    Shopping = "Shopping",
    Fitness = "Fitness",
    Coffee = "Coffee",
    Pets = "Pets",
    Tech = "Tech",
    Entertainment = "Entertainment",
    Music = "Music",
    Other = "Other",
    Chart = "Chart",
    SavingsGoal = "SavingsGoal",
    Exchange = "Exchange",
    CreditCard = "CreditCard",
    Currency = "Currency",
    Payments = "Payments",
    ATM = "ATM",
    Furniture = "Furniture",
    Maintenance = "Maintenance",
    Shipping = "Shipping",
    Internet = "Internet",
    Utilities = "Utilities",
    Idea = "Idea",
    Investment = "Investment",
    Company = "Company",
    Project = "Project",
    Team = "Team",
    Retail = "Retail",
    Art = "Art",
    Photography = "Photography",
    Reading = "Reading",
    Luxury = "Luxury",
    World = "World",
    Nature = "Nature",
    Subscription = "Subscription",
    Insurance = "Insurance",
    Podcast = "Podcast",
    Gadget = "Gadget",
    Donation = "Donation",
    Security = "Security",
    Gym = "Gym",
    Refund = "Refund",

    Clothes = "Clothes",
    Baby = "Baby",
    Beauty = "Beauty",
    Haircut = "Hair",
    Lost = "Lost",
    OtherIncome = "OtherIncome"
}